backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
jwt==1.3.1
email-validator
scikit-learn
pandas
openai==1.66.3
langchain
langchain-community
langchain-openai
chroma
chromadb
fastapi-mail